break up big ass main package
