+++ exporter/struct.go
@@ -1,14 +1,16 @@
-package main
+package exporter
 
 import "encoding/json"
 
 // Elasticsearch Node Stats Structs
 
+// NodeStatsResponse represents metrics from an Elasticsearch cluster
 type NodeStatsResponse struct {
 	ClusterName string `json:"cluster_name"`
 	Nodes       map[string]NodeStatsNodeResponse
 }
 
+// NodeStatsNodeResponse represents metrics about an Elasticsearch node
 type NodeStatsNodeResponse struct {
 	Name             string                                     `json:"name"`
 	Host             string                                     `json:"host"`
@@ -26,6 +28,7 @@ type NodeStatsNodeResponse struct {
 	Process          NodeStatsProcessResponse                   `json:"process"`
 }
 
+// NodeStatsBreakersResponse represents metrics about an Elasticsearch node breakers
 type NodeStatsBreakersResponse struct {
 	EstimatedSize int64   `json:"estimated_size_in_bytes"`
 	LimitSize     int64   `json:"limit_size_in_bytes"`
@@ -33,39 +36,47 @@ type NodeStatsBreakersResponse struct {
 	Tripped       int64   `json:"tripped"`
 }
 
+// NodeStatsJVMResponse represents JVM metrics for an Elasticsearch node
 type NodeStatsJVMResponse struct {
 	BufferPools map[string]NodeStatsJVMBufferPoolResponse `json:"buffer_pools"`
 	GC          NodeStatsJVMGCResponse                    `json:"gc"`
 	Mem         NodeStatsJVMMemResponse                   `json:"mem"`
 }
 
+// NodeStatsJVMGCResponse represents JVM GC metrics for an Elasticsearch node
 type NodeStatsJVMGCResponse struct {
 	Collectors map[string]NodeStatsJVMGCCollectorResponse `json:"collectors"`
 }
 
+// NodeStatsJVMGCCollectorResponse represents JVM GC collector metrics for an Elasticsearch node
 type NodeStatsJVMGCCollectorResponse struct {
 	CollectionCount int64 `json:"collection_count"`
 	CollectionTime  int64 `json:"collection_time_in_millis"`
 }
 
+// NodeStatsJVMBufferPoolResponse represents JVM buffer pool metrics for an Elasticsearch node
 type NodeStatsJVMBufferPoolResponse struct {
 	Count         int64 `json:"count"`
 	TotalCapacity int64 `json:"total_capacity_in_bytes"`
 	Used          int64 `json:"used_in_bytes"`
 }
 
+// NodeStatsJVMMemResponse represents JVM memory metrics for an Elasticsearch node
 type NodeStatsJVMMemResponse struct {
 	HeapCommitted    int64 `json:"heap_committed_in_bytes"`
 	HeapUsed         int64 `json:"heap_used_in_bytes"`
+	HeapUsedPercent  int64 `json:"heap_used_percent"`
 	HeapMax          int64 `json:"heap_max_in_bytes"`
 	NonHeapCommitted int64 `json:"non_heap_committed_in_bytes"`
 	NonHeapUsed      int64 `json:"non_heap_used_in_bytes"`
 }
 
+// NodeStatsNetworkResponse represents network metrics for an Elasticsearch node
 type NodeStatsNetworkResponse struct {
 	TCP NodeStatsTCPResponse `json:"tcp"`
 }
 
+// NodeStatsTransportResponse represents network transport metrics for an Elasticsearch node
 type NodeStatsTransportResponse struct {
 	ServerOpen int64 `json:"server_open"`
 	RxCount    int64 `json:"rx_count"`
@@ -74,6 +85,7 @@ type NodeStatsTransportResponse struct {
 	TxSize     int64 `json:"tx_size_in_bytes"`
 }
 
+// NodeStatsThreadPoolPoolResponse represents thread pool metrics for an Elasticsearch node
 type NodeStatsThreadPoolPoolResponse struct {
 	Threads   int64 `json:"threads"`
 	Queue     int64 `json:"queue"`
@@ -83,6 +95,7 @@ type NodeStatsThreadPoolPoolResponse struct {
 	Completed int64 `json:"completed"`
 }
 
+// NodeStatsTCPResponse represents TCP network metrics for an Elasticsearch node
 type NodeStatsTCPResponse struct {
 	ActiveOpens  int64 `json:"active_opens"`
 	PassiveOpens int64 `json:"passive_opens"`
@@ -96,6 +109,7 @@ type NodeStatsTCPResponse struct {
 	OutRsts      int64 `json:"out_rsts"`
 }
 
+// NodeStatsIndicesResponse represents index metrics for an Elasticsearch node
 type NodeStatsIndicesResponse struct {
 	Docs         NodeStatsIndicesDocsResponse
 	Store        NodeStatsIndicesStoreResponse
@@ -112,26 +126,31 @@ type NodeStatsIndicesResponse struct {
 	Refresh      NodeStatsIndicesRefreshResponse
 }
 
+// NodeStatsIndicesDocsResponse represents index document metrics for an Elasticsearch node
 type NodeStatsIndicesDocsResponse struct {
 	Count   int64 `json:"count"`
 	Deleted int64 `json:"deleted"`
 }
 
+// NodeStatsIndicesRefreshResponse represents index refresh metrics for an Elasticsearch node
 type NodeStatsIndicesRefreshResponse struct {
 	Total     int64 `json:"total"`
 	TotalTime int64 `json:"total_time_in_millis"`
 }
 
+// NodeStatsIndicesSegmentsResponse represents index segment metrics for an Elasticsearch node
 type NodeStatsIndicesSegmentsResponse struct {
 	Count  int64 `json:"count"`
 	Memory int64 `json:"memory_in_bytes"`
 }
 
+// NodeStatsIndicesStoreResponse represents index store metrics for an Elasticsearch node
 type NodeStatsIndicesStoreResponse struct {
 	Size         int64 `json:"size_in_bytes"`
 	ThrottleTime int64 `json:"throttle_time_in_millis"`
 }
 
+// NodeStatsIndicesIndexingResponse represents index metrics for an Elasticsearch node
 type NodeStatsIndicesIndexingResponse struct {
 	IndexTotal    int64 `json:"index_total"`
 	IndexTime     int64 `json:"index_time_in_millis"`
@@ -141,6 +160,7 @@ type NodeStatsIndicesIndexingResponse struct {
 	DeleteCurrent int64 `json:"delete_current"`
 }
 
+// NodeStatsIndicesMergesResponse represents index merge metrics for an Elasticsearch node
 type NodeStatsIndicesMergesResponse struct {
 	Current     int64 `json:"current"`
 	CurrentDocs int64 `json:"current_docs"`
@@ -151,6 +171,7 @@ type NodeStatsIndicesMergesResponse struct {
 	TotalTime   int64 `json:"total_time_in_millis"`
 }
 
+// NodeStatsIndicesGetResponse represents index get metrics for an Elasticsearch node
 type NodeStatsIndicesGetResponse struct {
 	Total        int64 `json:"total"`
 	Time         int64 `json:"time_in_millis"`
@@ -161,6 +182,7 @@ type NodeStatsIndicesGetResponse struct {
 	Current      int64 `json:"current"`
 }
 
+// NodeStatsIndicesSearchResponse represents index search metrics for an Elasticsearch node
 type NodeStatsIndicesSearchResponse struct {
 	OpenContext  int64 `json:"open_contexts"`
 	QueryTotal   int64 `json:"query_total"`
@@ -171,11 +193,13 @@ type NodeStatsIndicesSearchResponse struct {
 	FetchCurrent int64 `json:"fetch_current"`
 }
 
+// NodeStatsIndicesFlushResponse represents index flush metrics for an Elasticsearch node
 type NodeStatsIndicesFlushResponse struct {
 	Total int64 `json:"total"`
 	Time  int64 `json:"total_time_in_millis"`
 }
 
+// NodeStatsIndicesCacheResponse represents index cache metrics for an Elasticsearch node
 type NodeStatsIndicesCacheResponse struct {
 	Evictions  int64 `json:"evictions"`
 	MemorySize int64 `json:"memory_size_in_bytes"`
@@ -186,6 +210,7 @@ type NodeStatsIndicesCacheResponse struct {
 	TotalCount int64 `json:"total_count"`
 }
 
+// NodeStatsOSResponse represents OS metrics for an Elasticsearch node
 type NodeStatsOSResponse struct {
 	Timestamp int64 `json:"timestamp"`
 	Uptime    int64 `json:"uptime_in_millis"`
@@ -197,6 +222,7 @@ type NodeStatsOSResponse struct {
 	Swap    NodeStatsOSSwapResponse `json:"swap"`
 }
 
+// NodeStatsOSMemResponse represents OS memory metrics for an Elasticsearch node
 type NodeStatsOSMemResponse struct {
 	Free       int64 `json:"free_in_bytes"`
 	Used       int64 `json:"used_in_bytes"`
@@ -204,11 +230,13 @@ type NodeStatsOSMemResponse struct {
 	ActualUsed int64 `json:"actual_used_in_bytes"`
 }
 
+// NodeStatsOSSwapResponse represents OS swap metrics for an Elasticsearch node
 type NodeStatsOSSwapResponse struct {
 	Used int64 `json:"used_in_bytes"`
 	Free int64 `json:"free_in_bytes"`
 }
 
+// NodeStatsOSCPUResponse represents OS cpu metrics for an Elasticsearch node
 type NodeStatsOSCPUResponse struct {
 	Sys   int64 `json:"sys"`
 	User  int64 `json:"user"`
@@ -216,6 +244,7 @@ type NodeStatsOSCPUResponse struct {
 	Steal int64 `json:"stolen"`
 }
 
+// NodeStatsProcessResponse represents process metrics for an Elasticsearch node
 type NodeStatsProcessResponse struct {
 	Timestamp int64                       `json:"timestamp"`
 	OpenFD    int64                       `json:"open_file_descriptors"`
@@ -224,12 +253,14 @@ type NodeStatsProcessResponse struct {
 	Memory    NodeStatsProcessMemResponse `json:"mem"`
 }
 
+// NodeStatsProcessMemResponse represents process memory metrics for an Elasticsearch node
 type NodeStatsProcessMemResponse struct {
 	Resident     int64 `json:"resident_in_bytes"`
 	Share        int64 `json:"share_in_bytes"`
 	TotalVirtual int64 `json:"total_virtual_in_bytes"`
 }
 
+// NodeStatsProcessCPUResponse represents process cpu metrics for an Elasticsearch node
 type NodeStatsProcessCPUResponse struct {
 	Percent int64 `json:"percent"`
 	Sys     int64 `json:"sys_in_millis"`
@@ -237,16 +268,19 @@ type NodeStatsProcessCPUResponse struct {
 	Total   int64 `json:"total_in_millis"`
 }
 
+// NodeStatsHTTPResponse represents HTTP metrics for an Elasticsearch node
 type NodeStatsHTTPResponse struct {
 	CurrentOpen int64 `json:"current_open"`
 	TotalOpen   int64 `json:"total_open"`
 }
 
+// NodeStatsFSResponse represents filesystem metrics for an Elasticsearch node
 type NodeStatsFSResponse struct {
 	Timestamp int64                     `json:"timestamp"`
 	Data      []NodeStatsFSDataResponse `json:"data"`
 }
 
+// NodeStatsFSDataResponse represents filesystem data metrics for an Elasticsearch node
 type NodeStatsFSDataResponse struct {
 	Path          string `json:"path"`
 	Mount         string `json:"mount"`
